[Refactor] move transfer and mutating methods to separate classes
