+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2021-03-17 17:15:40 UTC using RuboCop version 1.11.0.
+# on 2021-03-17 17:24:07 UTC using RuboCop version 1.11.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
