+++ lib/dor/services/client/object.rb
@@ -4,7 +4,7 @@ module Dor
   module Services
     class Client
       # API calls that are about a repository object
-      class Object < VersionedService # rubocop:disable Metrics/ClassLength
+      class Object < VersionedService
         attr_reader :object_identifier
 
         # @param object_identifier [String] the pid for the object
@@ -73,25 +73,6 @@ module Dor
           [model, metadata]
         end
 
-        # Updates the object
-        # @param [Cocina::Models::RequestDRO,Cocina::Models::RequestCollection,Cocina::Models::RequestAPO] params model object
-        # @raise [NotFoundResponse] when the response is a 404 (object not found)
-        # @raise [UnexpectedResponse] when the response is not successful.
-        # @return [Cocina::Models::DRO,Cocina::Models::Collection,Cocina::Models::AdminPolicy] the returned model
-        def update(params:)
-          resp = connection.patch do |req|
-            req.url object_path
-            req.headers['Content-Type'] = 'application/json'
-            # asking the service to return JSON (else it'll be plain text)
-            req.headers['Accept'] = 'application/json'
-            req.body = params.to_json
-          end
-
-          return Cocina::Models.build(JSON.parse(resp.body)) if resp.success?
-
-          raise_exception_based_on_response!(resp)
-        end
-
         # Get a list of the collections. (Similar to Valkyrie's find_inverse_references_by)
         # @raise [UnexpectedResponse] if the request is unsuccessful.
         # @return [Array<Cocina::Models::DRO>]
@@ -106,56 +87,17 @@ module Dor
           Members.new(**parent_params).members
         end
 
-        # Publish an object (send to PURL)
-        # @raise [NotFoundResponse] when the response is a 404 (object not found)
-        # @raise [UnexpectedResponse] when the response is not successful.
-        # @param [String] workflow ('accessionWF') which workflow to callback to.
-        # @param [String] lane_id for prioritization (default or low)
-        # @return [boolean] true on success
-        def publish(workflow: nil, lane_id: nil)
-          query_params = [].tap do |params|
-            params << "workflow=#{workflow}" if workflow
-            params << "lane-id=#{lane_id}" if lane_id
-          end
-          query_string = query_params.any? ? "?#{query_params.join('&')}" : ''
-          publish_path = "#{object_path}/publish#{query_string}"
-          resp = connection.post do |req|
-            req.url publish_path
-          end
-          return resp.headers['Location'] if resp.success?
-
-          raise_exception_based_on_response!(resp)
+        def transfer
+          Transfer.new(**parent_params)
         end
 
-        # Preserve an object (send to SDR)
-        # @raise [NotFoundResponse] when the response is a 404 (object not found)
-        # @raise [UnexpectedResponse] when the response is not successful.
-        # @param [String] lane_id for prioritization (default or low)
-        # @return [String] URL from Location response header if no errors
-        def preserve(lane_id: nil)
-          query_string = lane_id ? "?lane-id=#{lane_id}" : ''
-          resp = connection.post do |req|
-            req.url "#{object_path}/preserve#{query_string}"
-          end
-          return resp.headers['Location'] if resp.success?
+        delegate :publish, :preserve, :shelve, to: :transfer
 
-          raise_exception_based_on_response!(resp)
+        def mutate
+          Mutate.new(**parent_params)
         end
 
-        # Shelve an object (send to Stacks)
-        # @raise [NotFoundResponse] when the response is a 404 (object not found)
-        # @raise [UnexpectedResponse] when the response is not successful.
-        # @param [String] lane_id for prioritization (default or low)
-        # @return [boolean] true on success
-        def shelve(lane_id: nil)
-          query_string = lane_id ? "?lane-id=#{lane_id}" : ''
-          resp = connection.post do |req|
-            req.url "#{object_path}/shelve#{query_string}"
-          end
-          return resp.headers['Location'] if resp.success?
-
-          raise_exception_based_on_response!(resp)
-        end
+        delegate :refresh_metadata, :update, :destroy, to: :mutate
 
         # Update the marc record for the given object
         # @raise [NotFoundResponse] when the response is a 404 (object not found)
@@ -168,32 +110,6 @@ module Dor
           return true if resp.success?
 
           raise_exception_based_on_response!(resp)
-        end
-
-        # Pull in metadata from Symphony and update descMetadata
-        # @raise [NotFoundResponse] when the response is a 404 (object not found)
-        # @raise [UnexpectedResponse] when the response is not successful.
-        # @return [boolean] true on success
-        def refresh_metadata
-          resp = connection.post do |req|
-            req.url "#{object_path}/refresh_metadata"
-          end
-          return true if resp.success?
-
-          raise_exception_based_on_response!(resp)
-        end
-
-        # Destroys an object
-        # @return [Boolean] true if successful
-        # @raise [NotFoundResponse] when the response is a 404 (object not found)
-        # @raise [UnexpectedResponse] if the request is unsuccessful.
-        def destroy
-          resp = connection.delete do |req|
-            req.url object_path
-          end
-          raise_exception_based_on_response!(resp, object_identifier) unless resp.success?
-
-          true
         end
 
         # Notify the external Goobi system for a new object that was registered in DOR
